Adding final count validation for gptransfer
Currently GPtransfer has count validation which runs after every table
copies data to the destination database. It does not do any checking at
the end of the transfer od data to the destination. Therefore, it does
not validate that the tables haven't changed in the meantime.

To resolve the issue, an additional row count validation has been added
with all the existing validation options. This count validation does a
final pass after the transfer is completely done to make sure the tables
haven't changed in the meantime. Therefore, the default is to compare
the row count of tables copied to the destination databases with the
tables in the source database.

An option '--no-final-count’ has beed introduced for users who do
not want to run the final count validation by default. The option
'--no-final-count’ will disable table row count validation that is
performed after GPtransfer completes copying database objects to the
target database.

Updated the GPtransfer help file with the description.